chore: refactor genai.TextModel for easier extension
- TextModel is now ConversationModel to reflect broader use cases than text
- modularized using strategy mixins to better separate functionality aspects
- previously stored models with KIND='genai.text' still work, however resolve to ConversationModel
- refactor genai modules into genai.retrieval, genai.models and genai.strategy
