+++ omegaml/backends/genai/models/__init__.py
@@ -1,0 +1,2 @@
+from .conversation import ConversationModel, ConversationModelBackend
+from .textmodel import TextModel, TextModelBackend
