+++ omegaml/backends/genai/retrieval/__init__.py
@@ -1,0 +1,5 @@
+from .embedding import SimpleEmbeddingModel
+from .index import DocumentIndex
+from .inmemory import InMemoryVectorStore
+from .mongovector import MongoDBVectorStore
+from .pgvector import PGVectorBackend
