+++ omegaml/server/tests/test_prompts.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from flask import url_for
 
-from omegaml.backends.genai.textmodel import TextModel
+from omegaml.backends.genai.models.conversation import ConversationModel
 from omegaml.server.app import create_app
 from omegaml.tests.util import OmegaTestMixin
 
@@ -23,36 +23,36 @@ class PromptsViewTest(OmegaTestMixin, TestCase):
     def test_prompt_save(self):
         om = self.om
         om.models.put('openai+http://localhost/mymodel', 'llms/mymodel',
-                      prompt='you are not very helpful')
+            prompt='you are not very helpful')
         # create a new prompt based on an existing model
         resp = self.client.post(url_for('omega-ai.prompts_api_save_prompt',
-                                        name='prompts/myprompt'),
-                                json={
-                                    'model': 'llms/mymodel',
-                                    'prompt': 'you are a very helpful assistant',
-                                })
+            name='prompts/myprompt'),
+            json={
+                'model': 'llms/mymodel',
+                'prompt': 'you are a very helpful assistant',
+            })
         self.assertEqual(resp.status_code, 200)
         self.assertIn('prompts/myprompt', om.models.list())
         meta = om.models.metadata('prompts/myprompt')
         self.assertEqual(meta.attributes['model'], 'llms/mymodel')
         self.assertEqual(meta.attributes['prompt'], 'you are a very helpful assistant')
         prompt = om.models.get('prompts/myprompt')
-        self.assertIsInstance(prompt, TextModel)
+        self.assertIsInstance(prompt, ConversationModel)
         self.assertEqual(prompt.model, 'mymodel')
         self.assertEqual(prompt.prompt, 'you are a very helpful assistant')
         # update the prompt
         resp = self.client.post(url_for('omega-ai.prompts_api_save_prompt',
-                                        name='prompts/myprompt'),
-                                json={
-                                    'model': 'llms/mymodel',
-                                    'prompt': 'you are a really very helpful assistant',
-                                })
+            name='prompts/myprompt'),
+            json={
+                'model': 'llms/mymodel',
+                'prompt': 'you are a really very helpful assistant',
+            })
         self.assertEqual(resp.status_code, 200)
         self.assertIn('prompts/myprompt', om.models.list())
         meta = om.models.metadata('prompts/myprompt')
         self.assertEqual(meta.attributes['model'], 'llms/mymodel')
         self.assertEqual(meta.attributes['prompt'], 'you are a really very helpful assistant')
         prompt = om.models.get('prompts/myprompt')
-        self.assertIsInstance(prompt, TextModel)
+        self.assertIsInstance(prompt, ConversationModel)
         self.assertEqual(prompt.model, 'mymodel')
         self.assertEqual(prompt.prompt, 'you are a really very helpful assistant')
